pax-app/Wiki#160 Message anc Chat Controllers checks if a key provided is valid regarding DB context

+++ src/app/controllers/ChatController.js
@@ -4,9 +4,15 @@ class ChatController {
   async index(req, res) {
     const { user_id, provider_id } = req.body;
 
+    if (user_id && provider_id)
+      return res.status(400).json({ error: 'Only one param required' });
+
     const chats = await Chat.findAll({
       where: user_id ? { user_id } : { provider_id },
     });
+
+    if (!chats.length)
+      return res.status(400).json({ error: 'Key provided not exists' });
 
     return res.json(chats);
   }
@@ -15,7 +21,6 @@ class ChatController {
     const { user_id, provider_id } = req.body;
 
     const chat = await Chat.create({ user_id, provider_id });
-
     return res.json({
       chat_id: chat.chat_id,
       user: chat.user_id,
@@ -27,9 +32,9 @@ class ChatController {
     const { chat_id } = req.params;
 
     const chat = await Chat.findByPk(chat_id);
+    if (!chat) return res.status(400).json({ error: 'Chat does not exists' });
 
     await Chat.destroy({ where: { chat_id } });
-
     return res.json(chat);
   }
 }
